Refactor ScalingFactorsGridSearch

+++ src/main/java/de/monticore/mlpipelines/automl/trainalgorithms/efficientnet/ScalingFactors.java
@@ -1,11 +1,11 @@
 package de.monticore.mlpipelines.automl.trainalgorithms.efficientnet;
 
 public class ScalingFactors {
-    public double alpha;
-    public double beta;
-    public double gamma;
+    public float alpha;
+    public float beta;
+    public float gamma;
 
-    public ScalingFactors(double alpha, double beta, double gamma) {
+    public ScalingFactors(float alpha, float beta, float gamma) {
         this.alpha = alpha;
         this.beta = beta;
         this.gamma = gamma;
